+++ src/Listener/SentryCommandListenerFactory.php
@@ -6,25 +6,38 @@ namespace Sirix\SentryPsr\Listener;
 
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
-use Psr\Container\NotFoundExceptionInterface;
 use Sentry\State\HubInterface;
+use Sirix\ContainerResolver\ConfigReader;
+use Sirix\ContainerResolver\ContainerResolver;
+use Sirix\Redaction\RedactorInterface;
+use Sirix\SentryPsr\Config\SentryPsrConfig;
 use Sirix\SentryPsr\Helper\LoggerHelper;
+use Sirix\SentryPsr\Lifecycle\SentryLifecycle;
+use Sirix\SentryPsr\Redaction\SentryRedactorFactory;
 
 class SentryCommandListenerFactory
 {
     /**
      * @throws ContainerExceptionInterface
-     * @throws NotFoundExceptionInterface
      */
     public function __invoke(ContainerInterface $container): SentryCommandListener
     {
-        $hub = $container->get(HubInterface::class);
+        $containerResolver = ContainerResolver::forFactory($container, self::class);
+        $configReader      = ConfigReader::fromContainer($containerResolver);
+        SentryPsrConfig::assertConfigured($configReader);
 
-        $logger = LoggerHelper::getLogger($container);
+        $redactor = $containerResolver->has(RedactorInterface::class)
+            ? $containerResolver->get(RedactorInterface::class)
+            : (new SentryRedactorFactory())($container);
 
         return new SentryCommandListener(
-            $hub,
-            $logger,
+            $containerResolver->get(HubInterface::class),
+            isolateScope: $configReader->requiredBool('sentry_psr.isolate_console_scope'),
+            flushOnTerminate: $configReader->requiredBool('sentry_psr.flush_on_console_terminate'),
+            captureConsoleInput: $configReader->requiredBool('sentry_psr.capture_console_input'),
+            logger: LoggerHelper::getLogger($containerResolver),
+            sentryLifecycle: $containerResolver->get(SentryLifecycle::class),
+            redactor: $redactor,
         );
     }
 }