+++ test/Middleware/SentryErrorMiddlewareFactoryTest.php
@@ -7,15 +7,14 @@ namespace Sirix\SentryPsr\Test\Middleware;
 use Monolog\Logger;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
-use Psr\Container\ContainerExceptionInterface;
-use Psr\Container\ContainerInterface;
-use Psr\Container\NotFoundExceptionInterface;
 use Psr\Log\LoggerInterface;
-use ReflectionException;
 use ReflectionProperty;
 use Sentry\State\HubInterface;
+use Sirix\SentryPsr\Lifecycle\SentryLifecycle;
 use Sirix\SentryPsr\Middleware\SentryErrorMiddleware;
 use Sirix\SentryPsr\Middleware\SentryErrorMiddlewareFactory;
+use Sirix\SentryPsr\Test\Config\SentryPsrConfigFixture;
+use Sirix\SentryPsr\Test\Container\InMemoryContainer;
 
 /**
  * @internal
@@ -23,93 +22,81 @@ use Sirix\SentryPsr\Middleware\SentryErrorMiddlewareFactory;
 #[CoversClass(SentryErrorMiddlewareFactory::class)]
 final class SentryErrorMiddlewareFactoryTest extends TestCase
 {
-    private ContainerInterface $containerMock;
     private HubInterface $hubMock;
+    private SentryLifecycle $lifecycle;
     private SentryErrorMiddlewareFactory $factory;
 
     public function setUp(): void
     {
         parent::setUp();
 
-        $this->containerMock = $this->createMock(ContainerInterface::class);
-        $this->hubMock       = $this->createMock(HubInterface::class);
-        $this->factory       = new SentryErrorMiddlewareFactory();
+        $this->hubMock   = $this->createMock(HubInterface::class);
+        $this->lifecycle = new SentryLifecycle($this->hubMock);
+        $this->factory   = new SentryErrorMiddlewareFactory();
     }
 
-    /**
-     * @throws ContainerExceptionInterface
-     * @throws NotFoundExceptionInterface
-     */
     public function testFactoryCreatesMiddlewareWithoutLogger(): void
     {
-        $this->containerMock->method('get')
-            ->with(HubInterface::class)
-            ->willReturn($this->hubMock)
-        ;
-        $this->containerMock->method('has')
-            ->willReturn(false)
-        ;
-
-        $middleware = $this->factory->__invoke($this->containerMock);
+        $middleware = $this->factory->__invoke(new InMemoryContainer([
+            HubInterface::class    => $this->hubMock,
+            SentryLifecycle::class => $this->lifecycle,
+            'config'               => SentryPsrConfigFixture::config(),
+        ]));
 
         $this->assertInstanceOf(SentryErrorMiddleware::class, $middleware);
     }
 
-    /**
-     * @throws ContainerExceptionInterface
-     * @throws ReflectionException
-     * @throws NotFoundExceptionInterface
-     */
     public function testFactoryCreatesMiddlewareWithLoggerInterface(): void
     {
         $logger = $this->createMock(LoggerInterface::class);
 
-        $this->containerMock->method('get')
-            ->willReturnMap([
-                [HubInterface::class, $this->hubMock],
-                [LoggerInterface::class, $logger],
-            ])
-        ;
-        $this->containerMock->method('has')
-            ->willReturnMap([
-                [LoggerInterface::class, true],
-                [Logger::class, false],
-                ['logger', false],
-            ])
-        ;
-
-        $middleware = $this->factory->__invoke($this->containerMock);
+        $middleware = $this->factory->__invoke(new InMemoryContainer([
+            HubInterface::class      => $this->hubMock,
+            SentryLifecycle::class   => $this->lifecycle,
+            LoggerInterface::class   => $logger,
+            'config'                 => SentryPsrConfigFixture::config(),
+        ]));
 
         $this->assertInstanceOf(SentryErrorMiddleware::class, $middleware);
         $this->assertSame($logger, (new ReflectionProperty($middleware, 'logger'))->getValue($middleware));
     }
 
-    /**
-     * @throws ContainerExceptionInterface
-     * @throws ReflectionException
-     * @throws NotFoundExceptionInterface
-     */
     public function testFactoryCreatesMiddlewareWithMonologLogger(): void
     {
         $monolog = $this->createMock(Logger::class);
 
-        $this->containerMock->method('get')
-            ->willReturnMap([
-                [HubInterface::class, $this->hubMock],
-                [Logger::class, $monolog],
-            ])
-        ;
-        $this->containerMock->method('has')
-            ->willReturnMap([
-                [LoggerInterface::class, false],
-                [Logger::class, true],
-                ['logger', false],
-            ])
-        ;
-
-        $middleware = $this->factory->__invoke($this->containerMock);
+        $middleware = $this->factory->__invoke(new InMemoryContainer([
+            HubInterface::class    => $this->hubMock,
+            SentryLifecycle::class => $this->lifecycle,
+            'Monolog\Logger'       => $monolog,
+            'config'               => SentryPsrConfigFixture::config(),
+        ]));
 
         $this->assertInstanceOf(SentryErrorMiddleware::class, $middleware);
         $this->assertSame($monolog, (new ReflectionProperty($middleware, 'logger'))->getValue($middleware));
+    }
+
+    public function testFactoryAppliesHttpConfigurationFlags(): void
+    {
+        $middleware = $this->factory->__invoke(new InMemoryContainer([
+            HubInterface::class    => $this->hubMock,
+            SentryLifecycle::class => $this->lifecycle,
+            'config'               => SentryPsrConfigFixture::config([
+                'isolate_http_scope'           => false,
+                'flush_on_http_error'          => true,
+                'capture_http_request_context' => true,
+                'http_context'                 => [
+                    'enabled'         => false,
+                    'capture_headers' => true,
+                ],
+            ]),
+        ]));
+
+        $this->assertFalse((new ReflectionProperty($middleware, 'isolateScope'))->getValue($middleware));
+        $this->assertTrue((new ReflectionProperty($middleware, 'flushOnHttpError'))->getValue($middleware));
+        $this->assertFalse((new ReflectionProperty($middleware, 'captureRequestContext'))->getValue($middleware));
+        $httpContext = (new ReflectionProperty($middleware, 'httpContext'))->getValue($middleware);
+        $this->assertFalse($httpContext['enabled']);
+        $this->assertTrue($httpContext['capture_headers']);
     }
 }